+++ src/lib/tcp/tests/mt_tcp_listener_mgr_unittests.cc
@@ -454,7 +454,7 @@ public:
         // Stop the listener and then verify it has stopped.
         ASSERT_NO_THROW_LOG(mt_listener_mgr_->stop());
         ASSERT_TRUE(mt_listener_mgr_->isStopped());
-        EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0);
+        EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0U);
 
         // Iterate over the clients, checking their outcomes.
         size_t total_responses = 0;
@@ -473,7 +473,7 @@ public:
             //
             // We expect 1 response.
             auto responses = client->getResponses();
-            ASSERT_EQ(responses.size(), 1);
+            ASSERT_EQ(responses.size(), 1U);
 
             // First we turn it into an Element tree.
             ConstElementPtr answer;
@@ -530,7 +530,8 @@ public:
 
         // Each thread-id ought to have handled the same number of clients.
         for (auto const& it : clients_per_thread) {
-            EXPECT_EQ(it.second, num_clients / clients_per_thread.size())
+            EXPECT_EQ(it.second,
+                      static_cast<long>(num_clients / clients_per_thread.size()))
                       << "thread-id: " << it.first
                       << ", clients: " << it.second << std::endl;
         }
@@ -628,7 +629,7 @@ public:
         // Stop the listener and then verify it has stopped.
         ASSERT_NO_THROW_LOG(mt_listener_mgr_->stop());
         ASSERT_TRUE(mt_listener_mgr_->isStopped());
-        EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0);
+        EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0U);
 
         // Iterate over the clients, checking their outcomes.
         size_t total_responses = 0;
@@ -648,7 +649,7 @@ public:
             //
             // We expect one response.
             auto responses = client->getResponses();
-            ASSERT_EQ(responses.size(), 1);
+            ASSERT_EQ(responses.size(), 1U);
 
             // First we turn it into an Element tree.
             ConstElementPtr answer;
@@ -692,7 +693,7 @@ public:
 
         // We should have had the expected number of pauses.
         if (!num_pauses) {
-            ASSERT_EQ(pause_cnt_, 0);
+            ASSERT_EQ(pause_cnt_, 0U);
         } else {
             // We allow a range on pauses of +-1.
             ASSERT_TRUE((num_pauses - 1) <= pause_cnt_ &&
@@ -774,14 +775,14 @@ TEST_F(MtTcpListenerMgrTest, basics) {
     // Verify the getters do what we expect.
     EXPECT_EQ(mt_listener_mgr_->getAddress(), address);
     EXPECT_EQ(mt_listener_mgr_->getPort(), port);
-    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 1);
+    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 1U);
     EXPECT_FALSE(mt_listener_mgr_->getTlsContext());
 
     // It should not have an IOService, should not be listening and
     // should have no threads.
     ASSERT_FALSE(mt_listener_mgr_->getThreadIOService());
     EXPECT_TRUE(mt_listener_mgr_->isStopped());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0U);
 
     // Verify that we cannot start it when multi-threading is disabled.
     ASSERT_FALSE(MultiThreadingMgr::instance().getMode());
@@ -791,7 +792,7 @@ TEST_F(MtTcpListenerMgrTest, basics) {
 
     // It should still not be listening and have no threads.
     EXPECT_TRUE(mt_listener_mgr_->isStopped());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0U);
 
     // Enable multi-threading.
     MultiThreadingMgr::instance().setMode(true);
@@ -799,7 +800,7 @@ TEST_F(MtTcpListenerMgrTest, basics) {
     // Make sure we can start it and it's listening with 1 thread.
     ASSERT_NO_THROW_LOG(mt_listener_mgr_->start());
     ASSERT_TRUE(mt_listener_mgr_->isRunning());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 1);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 1U);
     ASSERT_TRUE(mt_listener_mgr_->getThreadIOService());
     EXPECT_FALSE(mt_listener_mgr_->getThreadIOService()->stopped());
 
@@ -810,7 +811,7 @@ TEST_F(MtTcpListenerMgrTest, basics) {
     // Stop it and verify we're no longer listening.
     ASSERT_NO_THROW_LOG(mt_listener_mgr_->stop());
     ASSERT_TRUE(mt_listener_mgr_->isStopped());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0U);
     ASSERT_FALSE(mt_listener_mgr_->getThreadIOService());
 
     // Make sure we can call stop again without problems.
@@ -819,7 +820,7 @@ TEST_F(MtTcpListenerMgrTest, basics) {
     // We should be able to restart it.
     ASSERT_NO_THROW_LOG(mt_listener_mgr_->start());
     ASSERT_TRUE(mt_listener_mgr_->isRunning());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 1);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 1U);
     ASSERT_TRUE(mt_listener_mgr_->getThreadIOService());
     EXPECT_FALSE(mt_listener_mgr_->getThreadIOService()->stopped());
 
@@ -837,8 +838,8 @@ TEST_F(MtTcpListenerMgrTest, basics) {
     ASSERT_NO_THROW_LOG(mt_listener_mgr_->start());
     EXPECT_EQ(mt_listener_mgr_->getAddress(), address);
     EXPECT_EQ(mt_listener_mgr_->getPort(), port);
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 4);
-    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 4);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 4U);
+    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 4U);
     ASSERT_TRUE(mt_listener_mgr_->isRunning());
     ASSERT_TRUE(mt_listener_mgr_->getThreadIOService());
     EXPECT_FALSE(mt_listener_mgr_->getThreadIOService()->stopped());
@@ -847,24 +848,24 @@ TEST_F(MtTcpListenerMgrTest, basics) {
     // IOService stopped, state set to PAUSED.
     ASSERT_NO_THROW_LOG(mt_listener_mgr_->pause());
     ASSERT_TRUE(mt_listener_mgr_->isPaused());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 4);
-    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 4);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 4U);
+    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 4U);
     ASSERT_TRUE(mt_listener_mgr_->getThreadIOService());
     EXPECT_TRUE(mt_listener_mgr_->getThreadIOService()->stopped());
 
     // Verify we can resume it.
     ASSERT_NO_THROW_LOG(mt_listener_mgr_->resume());
     ASSERT_TRUE(mt_listener_mgr_->isRunning());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 4);
-    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 4);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 4U);
+    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 4U);
     ASSERT_TRUE(mt_listener_mgr_->getThreadIOService());
     EXPECT_FALSE(mt_listener_mgr_->getThreadIOService()->stopped());
 
     // Stop it and verify we're no longer listening.
     ASSERT_NO_THROW_LOG(mt_listener_mgr_->stop());
     ASSERT_TRUE(mt_listener_mgr_->isStopped());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0);
-    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 4);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0U);
+    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 4U);
     ASSERT_FALSE(mt_listener_mgr_->getThreadIOService());
     EXPECT_TRUE(mt_listener_mgr_->isStopped());
 }
@@ -941,23 +942,23 @@ TEST_F(MtTcpListenerMgrTest, tls) {
 
     EXPECT_EQ(mt_listener_mgr_->getAddress(), address);
     EXPECT_EQ(mt_listener_mgr_->getPort(), port);
-    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 1);
+    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 1U);
     EXPECT_EQ(mt_listener_mgr_->getTlsContext(), context);
     EXPECT_TRUE(mt_listener_mgr_->isStopped());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0U);
 
     // Make sure we can start it and it's listening with 1 thread.
     ASSERT_NO_THROW_LOG(mt_listener_mgr_->start());
     ASSERT_TRUE(mt_listener_mgr_->isRunning());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 1);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 1U);
     ASSERT_TRUE(mt_listener_mgr_->getThreadIOService());
     EXPECT_FALSE(mt_listener_mgr_->getThreadIOService()->stopped());
 
     // Stop it.
     ASSERT_NO_THROW_LOG(mt_listener_mgr_->stop());
     ASSERT_TRUE(mt_listener_mgr_->isStopped());
-    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0);
-    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 1);
+    EXPECT_EQ(mt_listener_mgr_->getThreadCount(), 0U);
+    EXPECT_EQ(mt_listener_mgr_->getThreadPoolSize(), 1U);
     ASSERT_FALSE(mt_listener_mgr_->getThreadIOService());
     EXPECT_TRUE(mt_listener_mgr_->isStopped());
 }
